+++ MekHQ/unittests/mekhq/campaign/finances/AccountantTest.java
@@ -52,8 +52,8 @@ import java.util.List;
 
 import megamek.common.Entity;
 import mekhq.campaign.Campaign;
-import mekhq.campaign.CampaignOptions;
 import mekhq.campaign.CurrentLocation;
+import mekhq.campaign.campaignOptions.CampaignOptions;
 import mekhq.campaign.personnel.Person;
 import mekhq.campaign.unit.Unit;
 import mekhq.campaign.universe.Faction;
